@@ -1,28 +1,32 @@
 ﻿using Microsoft.AspNetCore.Mvc.Testing;
-using Teapot.Web.Controllers;
 
 namespace Teapot.Web.Tests.IntegrationTests;
 
 [TestFixtureSource(typeof(HttpMethods), nameof(HttpMethods.All))]
-public class StatusCodeTests
+public class StatusCodeTests(HttpMethod httpMethod)
 {
-    private readonly HttpMethod _httpMethod;
-
-    private static readonly HttpClient _httpClient = new WebApplicationFactory<Program>().CreateDefaultClient();
-
-    public StatusCodeTests(HttpMethod httpMethod)
+    [OneTimeSetUp]
+    public void OneTimeSetUp()
     {
-        _httpMethod = httpMethod;
+        _httpClient = new WebApplicationFactory<Program>().CreateDefaultClient();
     }
+
+    [OneTimeTearDown]
+    public void OneTimeTearDown()
+    {
+        _httpClient.Dispose();
+    }
+
+    private HttpClient _httpClient = null!;
 
     [TestCaseSource(typeof(TestCases), nameof(TestCases.StatusCodesWithContent))]
     public async Task ResponseWithContent([Values] TestCase testCase)
     {
-        var uri = $"/{testCase.Code}";
-        using var httpRequest = new HttpRequestMessage(_httpMethod, uri);
-        using var response = await _httpClient.SendAsync(httpRequest);
+        string uri = $"/{testCase.Code}";
+        using HttpRequestMessage httpRequest = new(httpMethod, uri);
+        using HttpResponseMessage response = await _httpClient.SendAsync(httpRequest);
         Assert.That((int)response.StatusCode, Is.EqualTo(testCase.Code));
-        var body = await response.Content.ReadAsStringAsync();
+        string body = await response.Content.ReadAsStringAsync();
         Assert.Multiple(() =>
         {
             Assert.That(body.ReplaceLineEndings(), Is.EqualTo(testCase.Body));
@@ -35,11 +39,11 @@ public class StatusCodeTests
     [TestCaseSource(typeof(TestCases), nameof(TestCases.StatusCodesWithContent))]
     public async Task ResponseWithContentSuppressedViaQs([Values] TestCase testCase)
     {
-        var uri = $"/{testCase.Code}?{nameof(CustomHttpStatusCodeResult.SuppressBody)}=true";
-        using var httpRequest = new HttpRequestMessage(_httpMethod, uri);
-        using var response = await _httpClient.SendAsync(httpRequest);
+        string uri = $"/{testCase.Code}?{nameof(CustomHttpStatusCodeResult.SuppressBody)}=true";
+        using HttpRequestMessage httpRequest = new(httpMethod, uri);
+        using HttpResponseMessage response = await _httpClient.SendAsync(httpRequest);
         Assert.That((int)response.StatusCode, Is.EqualTo(testCase.Code));
-        var body = await response.Content.ReadAsStringAsync();
+        string body = await response.Content.ReadAsStringAsync();
         Assert.Multiple(() =>
         {
             Assert.That(body, Is.Empty);
@@ -52,12 +56,12 @@ public class StatusCodeTests
     [TestCaseSource(typeof(TestCases), nameof(TestCases.StatusCodesWithContent))]
     public async Task ResponseWithContentSuppressedViaHeader([Values] TestCase testCase)
     {
-        var uri = $"/{testCase.Code}";
-        using var httpRequest = new HttpRequestMessage(_httpMethod, uri);
-        httpRequest.Headers.Add(StatusController.SUPPRESS_BODY_HEADER, "true");
-        using var response = await _httpClient.SendAsync(httpRequest);
+        string uri = $"/{testCase.Code}";
+        using HttpRequestMessage httpRequest = new(httpMethod, uri);
+        httpRequest.Headers.Add(StatusExtensions.SUPPRESS_BODY_HEADER, "true");
+        using HttpResponseMessage response = await _httpClient.SendAsync(httpRequest);
         Assert.That((int)response.StatusCode, Is.EqualTo(testCase.Code));
-        var body = await response.Content.ReadAsStringAsync();
+        string body = await response.Content.ReadAsStringAsync();
         Assert.Multiple(() =>
         {
             Assert.That(body, Is.Empty);
@@ -70,11 +74,11 @@ public class StatusCodeTests
     [TestCaseSource(typeof(TestCases), nameof(TestCases.StatusCodesNoContent))]
     public async Task ResponseNoContent([Values] TestCase testCase)
     {
-        var uri = $"/{testCase.Code}";
-        using var httpRequest = new HttpRequestMessage(_httpMethod, uri);
-        using var response = await _httpClient.SendAsync(httpRequest);
+        string uri = $"/{testCase.Code}";
+        using HttpRequestMessage httpRequest = new(httpMethod, uri);
+        using HttpResponseMessage response = await _httpClient.SendAsync(httpRequest);
         Assert.That((int)response.StatusCode, Is.EqualTo(testCase.Code));
-        var body = await response.Content.ReadAsStringAsync();
+        string body = await response.Content.ReadAsStringAsync();
         Assert.Multiple(() =>
         {
             Assert.That(body, Is.Empty);
