@@ -1,6 +1,4 @@
 ﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Teapot.Web.Controllers;
 using Teapot.Web.Models;
 using Teapot.Web.Models.Unofficial;
 
@@ -8,30 +6,25 @@ namespace Teapot.Web.Tests.UnitTests;
 public class SleepTests {
     private const int Sleep = 500;
 
-    private TeapotStatusCodeResults _statusCodes;
+    private TeapotStatusCodeMetadataCollection _statusCodes;
 
     [SetUp]
     public void Setup() {
         _statusCodes = new(
-            new AmazonStatusCodeResults(),
-            new CloudflareStatusCodeResults(),
-            new EsriStatusCodeResults(),
-            new LaravelStatusCodeResults(),
-            new MicrosoftStatusCodeResults(),
-            new NginxStatusCodeResults(),
-            new TwitterStatusCodeResults()
+            new AmazonStatusCodeMetadata(),
+            new CloudflareStatusCodeMetadata(),
+            new EsriStatusCodeMetadata(),
+            new LaravelStatusCodeMetadata(),
+            new MicrosoftStatusCodeMetadata(),
+            new NginxStatusCodeMetadata(),
+            new TwitterStatusCodeMetadata()
         );
     }
 
     [Test]
     public void SleepReadFromQuery() {
-        StatusController controller = new(_statusCodes) {
-            ControllerContext = new ControllerContext {
-                HttpContext = new DefaultHttpContext()
-            }
-        };
-
-        IActionResult result = controller.StatusCode(200, Sleep, null);
+        Mock<HttpRequest> request = HttpRequestHelper.GenerateMockRequest();
+        IResult result = StatusExtensions.HandleStatusRequestAsync(200, Sleep, null, null, request.Object, _statusCodes);
 
         Assert.Multiple(() => {
             Assert.That(result, Is.InstanceOf<CustomHttpStatusCodeResult>());
@@ -42,15 +35,12 @@ public class SleepTests {
     }
 
     [Test]
-    public void SleepReadFromHeader() {
-        StatusController controller = new(_statusCodes) {
-            ControllerContext = new ControllerContext {
-                HttpContext = new DefaultHttpContext()
-            }
-        };
-        controller.ControllerContext.HttpContext.Request.Headers.Add(StatusController.SLEEP_HEADER, Sleep.ToString());
+    public void SleepReadFromHeader()
+    {
+        Mock<HttpRequest> request = HttpRequestHelper.GenerateMockRequest();
+        request.Object.Headers.Append(StatusExtensions.SLEEP_HEADER, Sleep.ToString());
 
-        IActionResult result = controller.StatusCode(200, null, null);
+        IResult result = StatusExtensions.HandleStatusRequestAsync(200, null, null, null, request.Object, _statusCodes);
 
         Assert.Multiple(() => {
             Assert.That(result, Is.InstanceOf<CustomHttpStatusCodeResult>());
@@ -62,14 +52,10 @@ public class SleepTests {
 
     [Test]
     public void SleepReadFromQSTakesPriorityHeader() {
-        StatusController controller = new(_statusCodes) {
-            ControllerContext = new ControllerContext {
-                HttpContext = new DefaultHttpContext()
-            }
-        };
-        controller.ControllerContext.HttpContext.Request.Headers.Add(StatusController.SLEEP_HEADER, Sleep.ToString());
+        Mock<HttpRequest> request = HttpRequestHelper.GenerateMockRequest();
+        request.Object.Headers.Append(StatusExtensions.SLEEP_HEADER, Sleep.ToString());
 
-        IActionResult result = controller.StatusCode(200, Sleep * 2, null);
+        IResult result = StatusExtensions.HandleStatusRequestAsync(200, Sleep * 2, null, null, request.Object, _statusCodes);
 
         Assert.Multiple(() => {
             Assert.That(result, Is.InstanceOf<CustomHttpStatusCodeResult>());
@@ -80,21 +66,20 @@ public class SleepTests {
     }
 
     [Test]
-    public void BadSleepHeaderIgnored() {
-        StatusController controller = new(_statusCodes) {
-            ControllerContext = new ControllerContext {
-                HttpContext = new DefaultHttpContext()
-            }
-        };
-        controller.ControllerContext.HttpContext.Request.Headers.Add(StatusController.SLEEP_HEADER, "invalid");
+    public void BadSleepHeaderIgnored()
+    {
+        Mock<HttpRequest> request = HttpRequestHelper.GenerateMockRequest();
+        request.Object.Headers.Append(StatusExtensions.SLEEP_HEADER, "invalid");
 
-        IActionResult result = controller.StatusCode(200, null, null);
+        IResult result = StatusExtensions.HandleStatusRequestAsync(200, null, null, null, request.Object, _statusCodes);
 
-        Assert.Multiple(() => {
+        Assert.Multiple(() =>
+        {
             Assert.That(result, Is.InstanceOf<CustomHttpStatusCodeResult>());
 
             CustomHttpStatusCodeResult r = (CustomHttpStatusCodeResult)result;
             Assert.That(r.Sleep, Is.Null);
         });
     }
+
 }
