@@ -1,23 +1,23 @@
 ﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Teapot.Web.Controllers;
 using Teapot.Web.Models;
 using Teapot.Web.Models.Unofficial;
 
 namespace Teapot.Web.Tests.UnitTests;
-public class SuppressBodyTests {
-    private TeapotStatusCodeResults _statusCodes;
+public class SuppressBodyTests
+{
+    private TeapotStatusCodeMetadataCollection _statusCodes;
 
     [SetUp]
-    public void Setup() {
+    public void Setup()
+    {
         _statusCodes = new(
-            new AmazonStatusCodeResults(),
-            new CloudflareStatusCodeResults(),
-            new EsriStatusCodeResults(),
-            new LaravelStatusCodeResults(),
-            new MicrosoftStatusCodeResults(),
-            new NginxStatusCodeResults(),
-            new TwitterStatusCodeResults()
+            new AmazonStatusCodeMetadata(),
+            new CloudflareStatusCodeMetadata(),
+            new EsriStatusCodeMetadata(),
+            new LaravelStatusCodeMetadata(),
+            new MicrosoftStatusCodeMetadata(),
+            new NginxStatusCodeMetadata(),
+            new TwitterStatusCodeMetadata()
         );
     }
 
@@ -25,18 +25,13 @@ public class SuppressBodyTests {
     [TestCase(true)]
     [TestCase(false)]
     [TestCase(null)]
-    public void SuppressBodyReadFromQuery(bool? suppressBody) {
-        StatusController controller = new(_statusCodes)
+    public void SuppressBodyReadFromQuery(bool? suppressBody)
+    {
+        Mock<HttpRequest> request = HttpRequestHelper.GenerateMockRequest();
+        IResult result = StatusExtensions.HandleStatusRequestAsync(200, null, suppressBody, null, request.Object, _statusCodes);
+
+        Assert.Multiple(() =>
         {
-            ControllerContext = new ControllerContext
-            {
-                HttpContext = new DefaultHttpContext()
-            }
-        };
-
-        IActionResult result = controller.StatusCode(200, null, suppressBody);
-
-        Assert.Multiple(() => {
             Assert.That(result, Is.InstanceOf<CustomHttpStatusCodeResult>());
 
             CustomHttpStatusCodeResult r = (CustomHttpStatusCodeResult)result;
@@ -50,25 +45,19 @@ public class SuppressBodyTests {
     [TestCase("")]
     public void SuppressBodyReadFromHeader(string? suppressBody)
     {
-        StatusController controller = new(_statusCodes)
-        {
-            ControllerContext = new ControllerContext
-            {
-                HttpContext = new DefaultHttpContext()
-            }
-        };
-        controller.ControllerContext.HttpContext.Request.Headers.Add(StatusController.SUPPRESS_BODY_HEADER, suppressBody);
+        Mock<HttpRequest> request = HttpRequestHelper.GenerateMockRequest();
+        request.Object.Headers.Append(StatusExtensions.SUPPRESS_BODY_HEADER, suppressBody);
 
-        IActionResult result = controller.StatusCode(200, null, null);
+        IResult result = StatusExtensions.HandleStatusRequestAsync(200, null, null, null, request.Object, _statusCodes);
 
         Assert.Multiple(() =>
         {
             Assert.That(result, Is.InstanceOf<CustomHttpStatusCodeResult>());
 
             CustomHttpStatusCodeResult r = (CustomHttpStatusCodeResult)result;
-            var expectedValue = suppressBody switch
+            bool? expectedValue = suppressBody switch
             {
-                string { Length: >0 } stringValue => (bool?)bool.Parse(stringValue),
+                string { Length: > 0 } stringValue => bool.Parse(stringValue),
                 _ => null
             };
             Assert.That(r.SuppressBody, Is.EqualTo(expectedValue));
@@ -84,25 +73,19 @@ public class SuppressBodyTests {
     [TestCase("", false)]
     public void SuppressBodyReadFromQSTakesPriorityHeader(string? headerValue, bool? queryStringValue)
     {
-        StatusController controller = new(_statusCodes)
-        {
-            ControllerContext = new ControllerContext
-            {
-                HttpContext = new DefaultHttpContext()
-            }
-        };
-        controller.ControllerContext.HttpContext.Request.Headers.Add(StatusController.SUPPRESS_BODY_HEADER, headerValue);
+        Mock<HttpRequest> request = HttpRequestHelper.GenerateMockRequest();
+        request.Object.Headers.Append(StatusExtensions.SUPPRESS_BODY_HEADER, headerValue);
 
-        IActionResult result = controller.StatusCode(200, null, queryStringValue);
+        IResult result = StatusExtensions.HandleStatusRequestAsync(200, null, queryStringValue, null, request.Object, _statusCodes);
 
         Assert.Multiple(() =>
         {
             Assert.That(result, Is.InstanceOf<CustomHttpStatusCodeResult>());
-            
+
             CustomHttpStatusCodeResult r = (CustomHttpStatusCodeResult)result;
-            var expectedValue = queryStringValue.HasValue ? queryStringValue.Value : headerValue switch
+            bool? expectedValue = queryStringValue ?? headerValue switch
             {
-                string { Length: > 0 } stringValue => (bool?)bool.Parse(stringValue),
+                string { Length: > 0 } stringValue => bool.Parse(stringValue),
                 _ => null
             };
             Assert.That(r.SuppressBody, Is.EqualTo(expectedValue));
@@ -112,16 +95,10 @@ public class SuppressBodyTests {
     [Test]
     public void BadSuppressBodyHeaderIgnored()
     {
-        StatusController controller = new(_statusCodes)
-        {
-            ControllerContext = new ControllerContext
-            {
-                HttpContext = new DefaultHttpContext()
-            }
-        };
-        controller.ControllerContext.HttpContext.Request.Headers.Add(StatusController.SUPPRESS_BODY_HEADER, "invalid");
+        Mock<HttpRequest> request = HttpRequestHelper.GenerateMockRequest();
+        request.Object.Headers.Append(StatusExtensions.SUPPRESS_BODY_HEADER, "invalid");
 
-        IActionResult result = controller.StatusCode(200, null, null);
+        IResult result = StatusExtensions.HandleStatusRequestAsync(200, null, null, null, request.Object, _statusCodes);
 
         Assert.Multiple(() =>
         {
