@@ -3,10 +3,10 @@ using System.Collections.Generic;
 
 namespace Teapot.Web.Models;
 
-public class TeapotStatusCodeResult
+public class TeapotStatusCodeMetadata
 {
     public string Description { get; set; } = "";
-    public Dictionary<string, string> IncludeHeaders { get; set; } = new();
+    public Dictionary<string, string> IncludeHeaders { get; set; } = [];
     public bool ExcludeBody { get; set; } = false;
     public Uri? Link { get; set; }
     public string? Body { get; set; }
