@@ -4,34 +4,34 @@ using Teapot.Web.Models.Unofficial;
 
 namespace Teapot.Web.Models;
 
-public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
+public class TeapotStatusCodeMetadataCollection : Dictionary<int, TeapotStatusCodeMetadata>
 {
-    public TeapotStatusCodeResults(
-        AmazonStatusCodeResults amazon,
-        CloudflareStatusCodeResults cloudflare,
-        EsriStatusCodeResults esri,
-        LaravelStatusCodeResults laravel,
-        MicrosoftStatusCodeResults microsoft,
-        NginxStatusCodeResults nginx,
-        TwitterStatusCodeResults twitter)
+    public TeapotStatusCodeMetadataCollection(
+        AmazonStatusCodeMetadata amazon,
+        CloudflareStatusCodeMetadata cloudflare,
+        EsriStatusCodeMetadata esri,
+        LaravelStatusCodeMetadata laravel,
+        MicrosoftStatusCodeMetadata microsoft,
+        NginxStatusCodeMetadata nginx,
+        TwitterStatusCodeMetadata twitter)
     {
         // 1xx range
-        Add(100, new TeapotStatusCodeResult
+        Add(100, new TeapotStatusCodeMetadata
         {
             Description = "Continue",
             ExcludeBody = true
         });
-        Add(101, new TeapotStatusCodeResult
+        Add(101, new TeapotStatusCodeMetadata
         {
             Description = "Switching Protocols",
             ExcludeBody = true
         });
-        Add(102, new TeapotStatusCodeResult
+        Add(102, new TeapotStatusCodeMetadata
         {
             Description = "Processing",
             ExcludeBody = true
         });
-        Add(103, new TeapotStatusCodeResult
+        Add(103, new TeapotStatusCodeMetadata
         {
             Description = "Early Hints",
             ExcludeBody = true,
@@ -42,33 +42,33 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
         });
 
         // 2xx range
-        Add(200, new TeapotStatusCodeResult
+        Add(200, new TeapotStatusCodeMetadata
         {
             Description = "OK",
         });
-        Add(201, new TeapotStatusCodeResult
+        Add(201, new TeapotStatusCodeMetadata
         {
             Description = "Created",
         });
-        Add(202, new TeapotStatusCodeResult
+        Add(202, new TeapotStatusCodeMetadata
         {
             Description = "Accepted"
         });
-        Add(203, new TeapotStatusCodeResult
+        Add(203, new TeapotStatusCodeMetadata
         {
             Description = "Non-Authoritative Information"
         });
-        Add(204, new TeapotStatusCodeResult
+        Add(204, new TeapotStatusCodeMetadata
         {
             Description = "No Content",
             ExcludeBody = true
         });
-        Add(205, new TeapotStatusCodeResult
+        Add(205, new TeapotStatusCodeMetadata
         {
             Description = "Reset Content",
             ExcludeBody = true
         });
-        Add(206, new TeapotStatusCodeResult
+        Add(206, new TeapotStatusCodeMetadata
         {
             Description = "Partial Content",
             IncludeHeaders = new Dictionary<string, string>
@@ -76,7 +76,7 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
                 {"Content-Range", "0-30"}
             }
         });
-        Add(207, new TeapotStatusCodeResult
+        Add(207, new TeapotStatusCodeMetadata
         {
             Description = "Multi-Status",
             IncludeHeaders = new Dictionary<string, string>
@@ -84,31 +84,33 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
                 { "Content-Type", "application/xml; charset=\"utf-8\"" }
             },
             Link = new Uri("https://tools.ietf.org/html/rfc4918"),
-            Body = @"<?xml version=""1.0"" encoding=""utf-8"" ?>
-<d:multistatus xmlns:d=""DAV:"">
-<d:response>
-    <d:href>http://www.example.com/container/resource3</d:href>
-    <d:status>HTTP/1.1 423 Locked</d:status>
-    <d:error><d:lock-token-submitted/></d:error>
-</d:response>
-</d:multistatus>"
+            Body = """
+            <?xml version="1.0" encoding="utf-8"?>
+            <d:multistatus xmlns:d="DAV:">
+            <d:response>
+                <d:href>http://www.example.com/container/resource3</d:href>
+                <d:status>HTTP/1.1 423 Locked</d:status>
+                <d:error><d:lock-token-submitted/></d:error>
+            </d:response>
+            </d:multistatus>
+            """
         });
-        Add(208, new TeapotStatusCodeResult
+        Add(208, new TeapotStatusCodeMetadata
         {
             Description = "Already Reported"
         });
-        Add(226, new TeapotStatusCodeResult
+        Add(226, new TeapotStatusCodeMetadata
         {
             Description = "IM Used",
             Link = new Uri("https://tools.ietf.org/html/rfc3229#section-10.4.1")
         });
 
         // 3xx range
-        Add(300, new TeapotStatusCodeResult
+        Add(300, new TeapotStatusCodeMetadata
         {
             Description = "Multiple Choices"
         });
-        Add(301, new TeapotStatusCodeResult
+        Add(301, new TeapotStatusCodeMetadata
         {
             Description = "Moved Permanently",
             IncludeHeaders = new Dictionary<string, string>
@@ -116,7 +118,7 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
                 {"Location", "https://httpstat.us"}
             }
         });
-        Add(302, new TeapotStatusCodeResult
+        Add(302, new TeapotStatusCodeMetadata
         {
             Description = "Found",
             IncludeHeaders = new Dictionary<string, string>
@@ -124,7 +126,7 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
                 {"Location", "https://httpstat.us"}
             }
         });
-        Add(303, new TeapotStatusCodeResult
+        Add(303, new TeapotStatusCodeMetadata
         {
             Description = "See Other",
             IncludeHeaders = new Dictionary<string, string>
@@ -132,12 +134,12 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
                 {"Location", "https://httpstat.us"}
             }
         });
-        Add(304, new TeapotStatusCodeResult
+        Add(304, new TeapotStatusCodeMetadata
         {
             Description = "Not Modified",
             ExcludeBody = true
         });
-        Add(305, new TeapotStatusCodeResult
+        Add(305, new TeapotStatusCodeMetadata
         {
             Description = "Use Proxy",
             IncludeHeaders = new Dictionary<string, string>
@@ -145,11 +147,11 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
                 {"Location", "https://httpstat.us"}
             }
         });
-        Add(306, new TeapotStatusCodeResult
+        Add(306, new TeapotStatusCodeMetadata
         {
             Description = "Switch Proxy"
         });
-        Add(307, new TeapotStatusCodeResult
+        Add(307, new TeapotStatusCodeMetadata
         {
             Description = "Temporary Redirect",
             IncludeHeaders = new Dictionary<string, string>
@@ -157,7 +159,7 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
                 {"Location", "https://httpstat.us"}
             }
         });
-        Add(308, new TeapotStatusCodeResult
+        Add(308, new TeapotStatusCodeMetadata
         {
             Description = "Permanent Redirect",
             IncludeHeaders = new Dictionary<string, string>
@@ -167,11 +169,11 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
         });
 
         // 4xx
-        Add(400, new TeapotStatusCodeResult
+        Add(400, new TeapotStatusCodeMetadata
         {
             Description = "Bad Request"
         });
-        Add(401, new TeapotStatusCodeResult
+        Add(401, new TeapotStatusCodeMetadata
         {
             Description = "Unauthorized",
             IncludeHeaders = new Dictionary<string, string>
@@ -179,27 +181,27 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
                 {"WWW-Authenticate", "Basic realm=\"Fake Realm\""}
             }
         });
-        Add(402, new TeapotStatusCodeResult
+        Add(402, new TeapotStatusCodeMetadata
         {
             Description = "Payment Required"
         });
-        Add(403, new TeapotStatusCodeResult
+        Add(403, new TeapotStatusCodeMetadata
         {
             Description = "Forbidden"
         });
-        Add(404, new TeapotStatusCodeResult
+        Add(404, new TeapotStatusCodeMetadata
         {
             Description = "Not Found"
         });
-        Add(405, new TeapotStatusCodeResult
+        Add(405, new TeapotStatusCodeMetadata
         {
             Description = "Method Not Allowed"
         });
-        Add(406, new TeapotStatusCodeResult
+        Add(406, new TeapotStatusCodeMetadata
         {
             Description = "Not Acceptable"
         });
-        Add(407, new TeapotStatusCodeResult
+        Add(407, new TeapotStatusCodeMetadata
         {
             Description = "Proxy Authentication Required",
             IncludeHeaders = new Dictionary<string, string>
@@ -207,80 +209,80 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
                 {"Proxy-Authenticate", "Basic realm=\"Fake Realm\""}
             }
         });
-        Add(408, new TeapotStatusCodeResult
+        Add(408, new TeapotStatusCodeMetadata
         {
             Description = "Request Timeout"
         });
-        Add(409, new TeapotStatusCodeResult
+        Add(409, new TeapotStatusCodeMetadata
         {
             Description = "Conflict"
         });
-        Add(410, new TeapotStatusCodeResult
+        Add(410, new TeapotStatusCodeMetadata
         {
             Description = "Gone"
         });
-        Add(411, new TeapotStatusCodeResult
+        Add(411, new TeapotStatusCodeMetadata
         {
             Description = "Length Required"
         });
-        Add(412, new TeapotStatusCodeResult
+        Add(412, new TeapotStatusCodeMetadata
         {
             Description = "Precondition Failed"
         });
-        Add(413, new TeapotStatusCodeResult
+        Add(413, new TeapotStatusCodeMetadata
         {
             Description = "Request Entity Too Large"
         });
-        Add(414, new TeapotStatusCodeResult
+        Add(414, new TeapotStatusCodeMetadata
         {
             Description = "Request-URI Too Long"
         });
-        Add(415, new TeapotStatusCodeResult
+        Add(415, new TeapotStatusCodeMetadata
         {
             Description = "Unsupported Media Type"
         });
-        Add(416, new TeapotStatusCodeResult
+        Add(416, new TeapotStatusCodeMetadata
         {
             Description = "Requested Range Not Satisfiable"
         });
-        Add(417, new TeapotStatusCodeResult
+        Add(417, new TeapotStatusCodeMetadata
         {
             Description = "Expectation Failed"
         });
-        Add(418, new TeapotStatusCodeResult
+        Add(418, new TeapotStatusCodeMetadata
         {
             Description = "I'm a teapot",
             Link = new Uri("https://www.ietf.org/rfc/rfc2324.txt")
         });
-        Add(421, new TeapotStatusCodeResult
+        Add(421, new TeapotStatusCodeMetadata
         {
             Description = "Misdirected Request"
         });
-        Add(422, new TeapotStatusCodeResult
+        Add(422, new TeapotStatusCodeMetadata
         {
             Description = "Unprocessable Entity"
         });
-        Add(423, new TeapotStatusCodeResult
+        Add(423, new TeapotStatusCodeMetadata
         {
             Description = "Locked"
         });
-        Add(424, new TeapotStatusCodeResult
+        Add(424, new TeapotStatusCodeMetadata
         {
             Description = "Failed Dependency"
         });
-        Add(425, new TeapotStatusCodeResult
+        Add(425, new TeapotStatusCodeMetadata
         {
             Description = "Too Early"
         });
-        Add(426, new TeapotStatusCodeResult
+        Add(426, new TeapotStatusCodeMetadata
         {
             Description = "Upgrade Required"
         });
-        Add(428, new TeapotStatusCodeResult
+        Add(428, new TeapotStatusCodeMetadata
         {
             Description = "Precondition Required"
         });
-        Add(429, new TeapotStatusCodeResult
+        Add(429, new TeapotStatusCodeMetadata
         {
             Description = "Too Many Requests",
             IncludeHeaders = new Dictionary<string, string>
@@ -288,59 +290,59 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
                 {"Retry-After", "5"}
             }
         });
-        Add(431, new TeapotStatusCodeResult
+        Add(431, new TeapotStatusCodeMetadata
         {
             Description = "Request Header Fields Too Large"
         });
-        Add(451, new TeapotStatusCodeResult
+        Add(451, new TeapotStatusCodeMetadata
         {
             Description = "Unavailable For Legal Reasons"
         });
 
         // 5xx
-        Add(500, new TeapotStatusCodeResult
+        Add(500, new TeapotStatusCodeMetadata
         {
             Description = "Internal Server Error"
         });
-        Add(501, new TeapotStatusCodeResult
+        Add(501, new TeapotStatusCodeMetadata
         {
             Description = "Not Implemented"
         });
-        Add(502, new TeapotStatusCodeResult
+        Add(502, new TeapotStatusCodeMetadata
         {
             Description = "Bad Gateway"
         });
-        Add(503, new TeapotStatusCodeResult
+        Add(503, new TeapotStatusCodeMetadata
         {
             Description = "Service Unavailable"
         });
-        Add(504, new TeapotStatusCodeResult
+        Add(504, new TeapotStatusCodeMetadata
         {
             Description = "Gateway Timeout"
         });
-        Add(505, new TeapotStatusCodeResult
+        Add(505, new TeapotStatusCodeMetadata
         {
             Description = "HTTP Version Not Supported"
         });
-        Add(506, new TeapotStatusCodeResult
+        Add(506, new TeapotStatusCodeMetadata
         {
             Description = "Variant Also Negotiates"
         });
-        Add(507, new TeapotStatusCodeResult
+        Add(507, new TeapotStatusCodeMetadata
         {
             Description = "Insufficient Storage"
         });
-        Add(508, new TeapotStatusCodeResult
+        Add(508, new TeapotStatusCodeMetadata
         {
             Description = "Loop Detected",
             Link = new Uri("https://tools.ietf.org/html/rfc5842")
         });
-        Add(510, new TeapotStatusCodeResult
+        Add(510, new TeapotStatusCodeMetadata
         {
             Description = "Not Extended",
             Link = new Uri("https://tools.ietf.org/html/rfc2774")
         });
-        Add(511, new TeapotStatusCodeResult
+        Add(511, new TeapotStatusCodeMetadata
         {
             Description = "Network Authentication Required"
         });
@@ -354,7 +356,7 @@ public class TeapotStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
         AddNonStandardStatusCodes(twitter);
     }
 
-    private void AddNonStandardStatusCodes(IDictionary<int, TeapotStatusCodeResult> codes)
+    private void AddNonStandardStatusCodes(IDictionary<int, TeapotStatusCodeMetadata> codes)
     {
         foreach (var item in codes)
         {
