@@ -2,11 +2,11 @@
 
 namespace Teapot.Web.Models.Unofficial;
 
-public class LaravelStatusCodeResults : Dictionary<int, TeapotStatusCodeResult>
+public class LaravelStatusCodeMetadata : Dictionary<int, TeapotStatusCodeMetadata>
 {
-    public LaravelStatusCodeResults()
+    public LaravelStatusCodeMetadata()
     {
-        Add(419, new TeapotStatusCodeResult
+        Add(419, new TeapotStatusCodeMetadata
         {
             Description = "CSRF Token Missing or Expired",
             IsNonStandard = true,
