@@ -1,4 +1,5 @@
-﻿<p>
+﻿@page
+<p>
     What you've requested doesn't exist, but here's a nice teapot for you to look at
     instead. :)
 </p>
